refactor: move command data structures to xlineapi
This commit simply do a move, it does not modify any existing code
